% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eilers_peeters_model.R
\name{ep_model}
\alias{ep_model}
\title{Fitting function for the EP model}
\usage{
ep_model(light, etr, a1 = 1e-06, b1 = 0.001, c1 = 3)
}
\arguments{
\item{light}{light values}

\item{etr}{etr values}

\item{b1}{}

\item{c1}{}
}
\value{
the fitted values for the EP model
}
\description{
Fitting function for the EP model
}
\keyword{external}
