% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eilers_peeters_model.R
\name{ep_model_simple}
\alias{ep_model_simple}
\title{Fitting function for the EP model (simplified)}
\usage{
ep_model_simple(light, etr, alpha = 0.1, Eopt = 300, Ps = 200)
}
\arguments{
\item{light}{light values}

\item{etr}{etr values}

\item{alpha}{}
}
\value{
the fitted values for the EP model
}
\description{
Fitting function for the EP model (simplified)
}
\keyword{external}
