% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscelaneous.R
\name{etr_calculations}
\alias{etr_calculations}
\title{Calculates ETR and rETR for each PAR}
\usage{
etr_calculations(fit_fluo, no_light_steps)
}
\value{
object with ETR and rETR
}
\description{
Calculates ETR and rETR for each PAR
}
\keyword{internal}
