% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_fluo.R
\name{fit_model_fofm}
\alias{fit_model_fofm}
\title{Function to import and fit PSI FL3500 data}
\usage{
fit_model_fofm(file_name, out_name = "results", samples = 1,
  num_steps = 11, sti = TRUE, str = TRUE, sti_model = "fofmsig",
  str_model = "tau2", calib_file = NA, protocol = "Campbell")
}
\arguments{
\item{sti}{fit the induction points, default = `true`}

\item{str}{fit the relaxation points, default = `true`}

\item{sti_model}{decide which sti model to fit c("fofmsig","fofmsigp","sigp","sig"). "fofmsigp", which fits fo fm sigmaPSII and rho; "fofmsig", which fits fo fm
and sigmaPSII and sets rho to 0; "sigp", which fits sigmaPSII and p, and calculates fo and fm as the mean of points specified in fo_pts and fm_pts; "sig", which fits sigmaPSII and sets p to 0, calcualtes fo and fm as the mean of points specified in fo_pts and fm_pts}

\item{str_model}{decide which str model to fit}

\item{protocol}{the current supported protocols are Campbell and RLC}
}
\value{
list all the objects the function returns plus the output files
}
\description{
Main processing scrip for PSI fluorometer generated files. Script cycles through all FluorWin output files in a specified folder and calculates FRR parameters.
}
