% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sti.R
\name{fit_sti}
\alias{fit_sti}
\title{Fitting function for the induction stage}
\usage{
fit_sti(x, y, fo.opt, fm.opt, sigma.opt, rho.opt, sigma_max, wm, sti.fit, c)
}
\arguments{
\item{x}{time}

\item{y}{fluorescence}

\item{fo.opt}{}

\item{rho.opt}{}

\item{sigma_max}{}

\item{wm}{}

\item{sti.fit}{}

\item{c}{is the object that has the sti data stored, it can have several columns
if there are several measurement}
}
\value{
a value that will be used in the fitting function
}
\description{
this will fit the Kolber et al. 1998 model to the induction step
}
\keyword{external}
