% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sti_simple.R
\name{fit_sti_simple}
\alias{fit_sti_simple}
\title{Fitting function for the induction stage}
\usage{
fit_sti_simple(x, y, fo.opt, fm.opt, sigma.opt, rho.opt, sigma_max, sti.fit)
}
\arguments{
\item{x}{time}

\item{y}{fluorescence}

\item{fo.opt}{}

\item{rho.opt}{}

\item{sigma_max}{}

\item{sti.fit}{}

\item{wm}{}
}
\value{
a value that will be used in the fitting function
}
\description{
this will fit the Kolber et al. 1998 model to the induction step
}
\keyword{external}
