% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_str.R
\name{fit_str}
\alias{fit_str}
\title{Fitting function for a FRRf re-opening curve}
\usage{
fit_str(x, y, fo = NULL, fm = NULL, tau_model = "tau1",
  subset_time = 500, plots = TRUE)
}
\arguments{
\item{x}{Vector with cumulative time in microseconds}

\item{y}{Fluorescence vector}

\item{fo}{Initial value for fo, if NULL then it defaults to the last point of the relaxation curve}

\item{fm}{Initial value for fm, if NULL then it defaults to the first point of the relaxation curve}

\item{tau_model}{Options: "tau1", "tau2", "tau3" and "tau1_subset" (the user selects a subset of the data starting from the first point of the re-opening curve). Default is tau1, fitting re-opening with a single phase exponential decay with lifetime of tau1.}

\item{subset_time}{Amount of time to fit the tau1_subset model (default = 500 us)}

\item{plots}{Turns on and off the plotting (TRUE, FALSE). Default is TRUE}
}
\value{
The function returns a list with: 1 - fitted parameters (list); 2 - fitted values (numeric vector); 3 - time in us (numeric vector)
}
\description{
This will fit a model to the re-opening phase using the equations of Zbigniew S. Kolber, Ondrej Prasil, Paul G. Falkowski (1998) Measurements of variable chlorophyll Fuorescence using fast repetition rate techniques: defining methodology and experimental protocols. Biochimica et Biophysica Acta, 1367: 88-106.
}
\examples{

fit_str(x = str_data$time, y = str_data$light_1,tau_model = 'tau1')

fit_str(x = str_data$time, y = str_data$light_1,tau_model = 'tau3')
}
\keyword{external}
