% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_cali.R
\docType{data}
\name{generic_cali}
\alias{generic_cali}
\title{Calibration file for package testing purposes}
\format{A csv file with a PSI FL-3500 calibration file}
\usage{
generic_cali
}
\description{
A csv file with a PSI FL-3500 calibration file for testing the psifluo
package. This is a typical output from a PSI FL-3500 calibration obtained using
Doug Campbell's protocol (see Vignette for more info).
}
\keyword{datasets}
