% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscelaneous.R
\name{initialize_constants}
\alias{initialize_constants}
\title{Initialize constants}
\usage{
initialize_constants()
}
\value{
list with the constants to use in the calculations
}
\description{
Creates default values for the constants
}
\keyword{internal}
