% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{kolber_sti}
\alias{kolber_sti}
\title{Model (Kolber et al. 1998) - induction}
\usage{
kolber_sti(incident, sigma, fo, fm, rho)
}
\arguments{
\item{incident}{flashlets, numeric}

\item{sigma}{value of sigma PSII}

\item{fo}{Fo value}

\item{fm}{Fm value}

\item{rho}{connectivity value (rho)}
}
\value{
a value that will be used in the fitting function
}
\description{
definition of the mode that will be fitted to the induction step
}
\keyword{internal}
