% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscelaneous.R
\name{npq_calculations}
\alias{npq_calculations}
\title{Calculate NPQ}
\usage{
npq_calculations(fit_fluo, no_light_steps)
}
\value{
object with NPQ
}
\description{
Calculate several parameters related to NPQ
}
\keyword{internal}
