% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npq_model.R
\name{npq_model}
\alias{npq_model}
\title{Fitting function for the EP model}
\usage{
npq_model(light, npq)
}
\arguments{
\item{light}{light values}

\item{npq}{npq values}
}
\value{
the parameters for the fitted model
}
\description{
Fitting function for the EP model
}
\keyword{external}
