% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscelaneous.R
\name{par_intensity}
\alias{par_intensity}
\title{Calculates the light intensity at each light level}
\usage{
par_intensity(protocol, no_light_steps, cali, nam)
}
\value{
PAR levels
}
\description{
Calculates the light intensity at each light level using the voltage values
}
\keyword{internal}
