% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_psi.R
\name{read_psi}
\alias{read_psi}
\title{Function to import and parse PSI text file}
\usage{
read_psi(file_name, out_name, samples, num_steps, calib_file)
}
\arguments{
\item{file_name}{name of the file to process, results will be written in this folder}

\item{out_name}{name for results files (default "results")}

\item{samples}{replicate names, if present should be in the form c("name1","name2")
otherwise leave at FALSE (default)}

\item{num_steps}{Manually input the number of light steps in each light curve (needs to be the same for all light curves in a data file). Defaults to 11, which is the maximum}

\item{calib_file}{select the calibration file}
}
\value{
list all the objects the function returns plus the output files
}
\description{
the result from this function should be an object that will be used by the fitting functions
}
