% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_psi_campbell.R
\name{read_psi_file_campbell}
\alias{read_psi_file_campbell}
\title{Function to import data from PSI FRRF using Doug Campbell double light/dark FRRf protocol}
\usage{
read_psi_file_campbell(filename, calib_file = NA, dec = ".")
}
\arguments{
\item{filename}{Path to the file to be imported}

\item{calib_file}{Path to the calibration file to convert lamp voltage settings to photons. A^-2^ . s^-1^. If not provided the default is a calibration function from a Doug Campbell PSI FL3500 instrument. If you are not using his machine the parameters sigmaPSII, cumulative excitation and actinic PAR values associated to an individual machine will be wrongly calibrated.}

\item{dec}{Decimal separator, default= "."}
}
\value{
Returns a list of five data frames and one numeric vector. 1 - all.data  = raw data input; 2 - meta.data = acquisiton settings; flashlet_energy = the calculated flashlet energy from the calibration file; 3 - data.sti  = induction only; 4 - data.str  = re-opening only; 5 - measuring_steps = voltage and equivalent actinic PAR at each step.
}
\description{
Imports a PSI FRRF data file generated by applying an FRRf induction/relaxation on top of a level of actinic light (which may  = 0)  and then again after a user defined dark period (ex. 1 s) to allow re-opening of PSII. Reads in data file, parses metadata, subtracts electronic background from fluorescence signal during light flashlets, splits columns into "light" and "dark" vectors,  splits data by sti and str, calculates the actinic light level at each step.
}
\keyword{external}
