% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{subset_str}
\alias{subset_str}
\title{Simple negative exponential to a subset of the data in the relaxation curve}
\usage{
subset_str(x, fo, fm, tau1)
}
\arguments{
\item{x}{Time}

\item{fo}{Fo value}

\item{fm}{Fm value}

\item{tau1}{Parameter that will store the value of tau1}
}
\value{
A numerical value that will be used in the fitting function
}
\description{
Definition of the model that will be fitted to the relaxation step
}
\keyword{internal}
